esp32c3: Implement events
Events detect changes in the states of device sensors.

Two types of events have been implemented: interrupt events,
which are transmitted to a broker as soon as a sensor state change
occurs, and periodic events, which are transmitted at regular time
intervals.

The events can be boolean for binary sensors or have a specific value
type for numeric sensors.