controller: Add APIs to interact with device events
Add APIs to fetch events sent by a device to a broker, with the
broker address retrieved from the device description

Among all APIs, two provides the core functionalities for events.

The first API starts an asynchronous task for each device that supports
events. A single receiver is returned to aggregate all the events coming
from the different device brokers. Events can then be parsed and
filtered by their device IDs. This API has been mainly thought for
TUI interfaces.

The second API starts an asynchronous task for a device that
supports events, interacting with the broker in the same way as the
first API. It is primarily designed for an SSE model.

There is also an API to stop the asynchronous tasks and clean up all
used resources.